refactor: move some gumby out of CW files
We are moving gumby commands since this causes CW
to import gumby, and gumby code uses fs-extra, and
this breaks browser mode
